requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain-community
langchain-openai
openai
tqdm
